Consolidate editorial widget view code
